Adding Order success page

+++ src/redux/slices/categorySlice.jsx
@@ -12,11 +12,25 @@ export const fetchCategories = createAsyncThunk('category/fetchAll', async (_, t
         return thunkAPI.rejectWithValue(error.response?.data || 'Fetch failed');
     }
 });
+export const fetchCategoryById = createAsyncThunk(
+    'category/fetchById',
+    async (id, thunkAPI) => {
+        try {
+            const res = await axios.get(`${APP_URL}/categories/${id}`);
+            return res.data.data;
+        } catch (error) {
+            return thunkAPI.rejectWithValue(error.response?.data || 'Fetch failed');
+        }
+    }
+);
 export const createCategory = createAsyncThunk(
     'category/create',
     async (formData, thunkAPI) => {
         try {
             const token = localStorage.getItem('token');
+            if (!token || token == "undefined") {
+                return thunkAPI.rejectWithValue('No token found');
+            }
 
             const res = await axios.post(`${APP_URL}/categories`, formData, {
                 headers: {
@@ -31,12 +45,37 @@ export const createCategory = createAsyncThunk(
         }
     }
 );
+export const updateCategory = createAsyncThunk(
+    'category/update',
+    async ({ id, formData }, thunkAPI) => {
+        try {
+            const token = localStorage.getItem('token');
+            if (!token || token === "undefined") {
+                return thunkAPI.rejectWithValue('No token found');
+            }
+
+            const res = await axios.put(`${APP_URL}/categories/${id}?_method=PUT`, formData, {
+                headers: {
+                    Authorization: `Bearer ${token}`,
+                    'Content-Type': 'multipart/form-data',
+                },
+            });
+
+            return res.data.data;
+        } catch (error) {
+            return thunkAPI.rejectWithValue(error.response?.data || 'Update failed');
+        }
+    }
+);
 
 export const deleteCategory = createAsyncThunk(
     'category/delete',
     async (id, thunkAPI) => {
         try {
             const token = localStorage.getItem('token');
+            if (!token || token == "undefined") {
+                return thunkAPI.rejectWithValue('No token found');
+            }
 
             await axios.delete(`${APP_URL}/categories/${id}`, {
                 headers: {
@@ -70,6 +109,17 @@ const categorySlice = createSlice({
                 state.error = action.payload;
                 state.loading = false;
             })
+            .addCase(fetchCategoryById.pending, (state) => {
+                state.loading = true;
+            })
+            .addCase(fetchCategoryById.fulfilled, (state, action) => {
+                state.loading = false;
+                state.selectedCategory = action.payload;
+            })
+            .addCase(fetchCategoryById.rejected, (state, action) => {
+                state.loading = false;
+                state.error = action.payload;
+            })
             .addCase(createCategory.pending, (state) => {
                 state.loading = true;
             })
@@ -80,6 +130,17 @@ const categorySlice = createSlice({
             .addCase(createCategory.rejected, (state, action) => {
                 state.error = action.payload;
                 state.loading = false;
+            })
+            .addCase(updateCategory.pending, (state) => {
+                state.loading = true;
+            })
+            .addCase(updateCategory.fulfilled, (state, action) => {
+                state.loading = false;
+                state.selectedCategory = action.payload;
+            })
+            .addCase(updateCategory.rejected, (state, action) => {
+                state.loading = false;
+                state.error = action.payload;
             })
             .addCase(deleteCategory.pending, (state) => {
                 state.loading = true;
